SDK_LEARNING/LEARNING_WITH_ONLINE_CLASS/PRACTICE_CODE/INPUT_GUARDRAIL/pyproject.toml
[project]
name = "guardrail"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.0.16",
    "python-dotenv>=1.1.0",
]
